Move modal grid layout to common place
