+++ client/packages/common/src/ui/layout/skeletons/index.ts
@@ -1,3 +1,4 @@
 export * from './DataTableSkeleton';
 export * from './DetailViewSkeleton';
 export * from './DetailFormSkeleton';
+export * from './ModalGridLayout';
